Added peak-phase-analysis code
Current version uses Offelli and Petri algorithm to compute an accurate value of phase for a given Fourier peak. This should be more accurate than just taking the phase of the (discrete) Fourier peak, but the OP approach is oddly terraced, so I've added a version which just uses the peak phase.